+++ docs/make_docs.sh
@@ -6,11 +6,10 @@ MAKE_DOCS_PATH="$( cd "$(dirname "$0")" ; pwd -P )"
 popd > /dev/null
 
 # generate the ReST files
-echo "debug"
-echo ${MAKE_DOCS_PATH}/../zfit_physics
+#echo ${MAKE_DOCS_PATH}/../zfit_physics
 ls ${MAKE_DOCS_PATH}
 ls .
-sphinx-apidoc -o ${MAKE_DOCS_PATH}/api ${MAKE_DOCS_PATH}/../zfit_physics -fMeT && \
+sphinx-apidoc -o ${MAKE_DOCS_PATH}/api ${MAKE_DOCS_PATH}/../src/zfit_physics -fMeT && \
 python3 ${MAKE_DOCS_PATH}/api/tools/change_headline.py ${MAKE_DOCS_PATH}/api/zfit_physics.* && \
 make -C ${MAKE_DOCS_PATH} clean && make -C ${MAKE_DOCS_PATH} html -j8 && \
 echo "Documentation successfully built!" || echo "FAILED to build Documentation"