+++ src/zfit_physics/compwa/__init__.py
@@ -1,0 +1,3 @@
+from . import data, loss, pdf, variables
+
+__all__ = ["pdf", "variables", "loss"]