+++ src/zfit_physics/pdf.py
@@ -8,4 +8,4 @@ from .models.pdf_novosibirsk import Novosibirsk
 from .models.pdf_relbw import RelativisticBreitWigner
 from .models.pdf_tsallis import Tsallis
 
-__all__ = ["Argus", "RelativisticBreitWigner", "CMSShape", "Cruijff", "ErfExp", "Novosibirsk", "Tsallis"]
+__all__ = ["Argus", "CMSShape", "Cruijff", "ErfExp", "Novosibirsk", "RelativisticBreitWigner", "Tsallis"]