+++ tests/test_pdf_relbw.py
@@ -34,14 +34,14 @@ def test_relbw_pdf():
 def test_relbw_integral():
     # Test CDF and integral here
     relbw, obs = create_relbw(m_true, gamma_true, limits=(0, 200))
-    full_interval_analytic = zfit.run(relbw.analytic_integrate(obs, norm_range=False))
-    full_interval_numeric = zfit.run(relbw.numeric_integrate(obs, norm_range=False))
+    full_interval_analytic = zfit.run(relbw.analytic_integrate(obs, norm=False))
+    full_interval_numeric = zfit.run(relbw.numeric_integrate(obs, norm=False))
     true_integral = 0.99888
     assert full_interval_analytic == pytest.approx(true_integral, 1e-4)
     assert full_interval_numeric == pytest.approx(true_integral, 1e-2)
 
-    analytic_integral = zfit.run(relbw.analytic_integrate(limits=(50, 100), norm_range=False))
-    numeric_integral = zfit.run(relbw.numeric_integrate(limits=(50, 100), norm_range=False))
+    analytic_integral = zfit.run(relbw.analytic_integrate(limits=(50, 100), norm=False))
+    numeric_integral = zfit.run(relbw.numeric_integrate(limits=(50, 100), norm=False))
     assert analytic_integral == pytest.approx(numeric_integral, 0.01)